+++ styleguide/haml_locals.rb
@@ -1,23 +1,25 @@
+# frozen_string_literal: true
+
 @locals = {
   breadcrumb_links: [
-      {
-          url: "https://www.citizensadvice.org.uk/benefits/",
-          label: "Benefits"
-      },
-      {
-          url: "https://www.citizensadvice.org.uk/benefits/benefits-introduction/",
-          label: "Benefits - introduction"
-      },
-      {
-          url: "https://www.citizensadvice.org.uk/benefits/benefits-introduction/what-benefits-can-i-get/",
-          label: "Benefit calculators: what benefits can you get"
-      }
+    {
+      url: 'https://www.citizensadvice.org.uk/benefits/',
+      label: 'Benefits'
+    },
+    {
+      url: 'https://www.citizensadvice.org.uk/benefits/benefits-introduction/',
+      label: 'Benefits - introduction'
+    },
+    {
+      url: 'https://www.citizensadvice.org.uk/benefits/benefits-introduction/what-benefits-can-i-get/',
+      label: 'Benefit calculators: what benefits can you get'
+    }
   ],
 
   notice: {
-    type: "warning",
-    title: "I am the title",
-    body: "The Body"
+    type: 'warning',
+    title: 'I am the title',
+    body: 'The Body'
   },
 
   root_path: 'root_path',
@@ -25,14 +27,19 @@
   sign_in_url: '/sign-in/url',
   search_action_url: '/the/search-action-url',
 
-  notice_banner_content: "If you’re a Thomas Cook customer and you’re stuck abroad or want to get your money back, get help from the Civil Aviation Authority.",
+  notice_banner_content: 'If you’re a Thomas Cook customer and you’re stuck abroad or want to get your money back, get help from the Civil Aviation Authority.',
 
-  radio_group_name: "radiogroup-1",
-  radio_group_label: "Select an option",
-  radio_group_hint: "There are two options",
-  radio_group_options: ["Option 1", "Option 2"],
+  radio_group_name: 'radiogroup-1',
+  radio_group_label: 'Select an option',
+  radio_group_hint: 'There are two options',
+  radio_group_options: ['Option 1', 'Option 2'],
 
-  input_name: "inputABC",
-  input_label: "This is the label for the input",
-  input_hint: "This is the hint for the input"
+  input_name: 'inputABC',
+  input_label: 'This is the label for the input',
+  input_hint: 'This is the hint for the input',
+
+  contact_details: {
+    title: 'Bail for Immigration Detainees',
+    body: "<p>Telephone: 01234 567890</p><p>Monday to Thursday 8am to 1pm</p><p>Calls cost a lot of money</p><p><a href='http://link.to.the.website'>Link to the website</a></p>"
+  }
 }